+++ changelog.md
@@ -12,6 +12,7 @@ All notable changes to the BSPED Paediatric DKA Calculator codebase will be docu
 
 ### Added
 
+- There is now a facility to update limited episode data after submission. This is currently limited to preventable factors and pre-existing diabetes status, but may be expanded in the future to allow retrospective audit data collection.
 - A unique patient hash is generated and stored using SHA-256 (x1 client-side, x1 server-side with salt) using patient NHS number and date of birth as input string. This allows episodes relating to the same patient to be linked for audit purposes.
 - The patient postcode is collected and sent with other data to the API where an index of multiple deprivation (IMD) decile is derived and stored. The postcode is not stored.
 - Bicarbonate, glucose and ketones may now optionally be entered and will be printed on the care pathway if provided. Bicarbonate is used for severity scoring as above.
@@ -20,7 +21,6 @@ All notable changes to the BSPED Paediatric DKA Calculator codebase will be docu
 - The insulin delivery method the patient uses (if they have pre-existing diabetes) is now collected for audit purposes.
 - Factors that may have contributed to the episode of DKA occuring (preventable factors) are now collected for audit purposes.
 - When the patient has pre-existing diabetes ticks are automatically added to the not applicable boxes for ongoing insulin therapy for the mode the patient does not use on flow chart 5.
-- There is now a facility to update limited episode data after submission. This is currently limited to preventable factors and pre-existing diabetes status, but may be expanded in the future to allow retrospective audit data collection.
 
 ## [v1.3.5] - 2022-12-06 08:55